Removed unused methods + Refactore file to meaningul name + Updated README
